Feature04/user score update, receive and send review

+++ src/main/java/com/beteam/willu/review/dto/ReviewRequestDto.java
@@ -1,11 +1,12 @@
 package com.beteam.willu.review.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
+
 import lombok.Getter;
 
 @Getter
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class ReviewRequestDto {
-    private String content;
-    private long score;
+	private String content;
+	private Double score;
 }
